requirements.txt
python-dotenv>=1.0.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
chardet>=5.0.0
tqdm>=4.65.0
streamlit>=1.32.0
plotly>=5.18.0
watchdog>=3.0.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.120.0
